Added datetime of docs as a time_t number (also updates perl module) Fixed Win32 problem as reported by David Norris MAXHITS removed Updated getent in html.c to improve performance Updated numerical compress macros to faster ones as suggested by Rainer Updated read_stream for faster filetring (fgetc changed by fread) AsciiEntities yes|no in config file Added -P switch to delimit the phrase search

+++ src/check.c
@@ -175,7 +175,7 @@ int  lchecksuffix;
 
    while (fl != NULL) {
         if (lstrstr(fl->suffix, checksuffix)
-            && strlen(fl->suffix) == lchecksuffix) {
+            && ((int)strlen(fl->suffix)) == lchecksuffix) {
 		efree(checksuffix);
 		return fl->prog;
 	}
@@ -206,7 +206,7 @@ struct swline *swl;
 		for(swl=ic->patt;swl;swl=swl->next)
 		{
 			if (lstrstr(swl->line, checksuffix)
-			   && strlen(swl->line) == lchecksuffix) {
+			   && ((int)strlen(swl->line)) == lchecksuffix) {
 				efree(checksuffix);
 				return ic->DocType;
 			}
